+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Added
+- Imperial units
+- NaN, PositiveInfinity and NegativeInfinity quantities (#83)
+- LinearMassDensity (#82)
+- Scandinavian mile length unit (#88)
+- Custom length example (#89)
+
+### Fixed
+- null reference exception in QuantityJsonConverter (#85)
+
 ## 3.0.1
 ### Changed
 - IQuantity depends on IXmlSerializable - changed XML serialization  (#62)
@@ -14,8 +25,6 @@ All notable changes to this project will be documented in this file.
 - TimeSpan extension methods (#43)
 - Format with conversion, without showing unit (#52)
 - Moment type (#11)
-- Imperial units
-- NaN, PositiveInfinity and NegativeInfinity quantities (#83)
 
 ### Fixed
 - lb/gal unit (#1)